requirements.txt
autogluon==1.0.0
autogluon.common==1.0.0
autogluon.core==1.0.0
autogluon.features==1.0.0
autogluon.multimodal==1.0.0
autogluon.tabular==1.0.0
autogluon.timeseries==1.0.0
gdown==5.1.0
MAPIE==0.8.2
numpy==1.26.4
pandas==2.2.1
Pillow==10.2.0
scikit_learn==1.4.1.post1
torch==2.0.0.post304
tqdm==4.66.2
